rename bambu
