+++ siliconcompiler/tools/__init__.py
@@ -1,4 +1,4 @@
-from siliconcompiler.tools.bambu import bambu
+from siliconcompiler.tools import bambu
 from siliconcompiler.tools.bluespec import bluespec
 from siliconcompiler.tools.builtin import builtin
 from siliconcompiler.tools.chisel import chisel
